Convert simple Audio to new structure
No more AudioHOC
