+++ examples/audio_classification/completions/1.json
@@ -18,7 +18,7 @@
         }
     ],
     "data": {
-        "url": "https://go.heartex.net/static/samples/game.wav"
+        "url": "https://app.heartex.net/static/samples/game.wav"
     },
     "id": 1,
     "task_path": "../examples/audio_classification/tasks.json"
