+++ examples/audio_classification/tasks.json
@@ -1,7 +1,7 @@
 [
     {
         "data": {
-            "url": "https://go.heartex.net/static/samples/game.wav"
+            "url": "https://app.heartex.net/static/samples/game.wav"
         },
         "predictions": [
             {
@@ -25,7 +25,7 @@
     },
     {
         "data": {
-            "url": "https://go.heartex.net/static/samples/game.wav"
+            "url": "https://app.heartex.net/static/samples/game.wav"
         }
     }
 ]
